🗑️ chore(mcp): remove API MCP module and decouple trial flow
Delete the organizations/mcp NestJS module (controller/service/module) and
its RouterModule wiring in app.module and organizations.module.

The public trial (quick-start) flow reused McpService only to sign a JWT for
the trial user. Replace that dependency with a self-contained TrialTokenService
(its own JwtModule, same secret env for token compatibility). Rename the trial
response/body fields from mcpToken/mcpUrl to trialToken and drop the MCP URL.

Also remove the now-unused MCP type contracts and ports from @packmind/types
(IGetMcpTokenUseCase, IGetMcpUrlUseCase, IMcpTokenService).

Part of dropping the MCP feature (#651).

(cherry picked from commit 5246addf8cac601ec3a773284cd02c61f55c4b92)
