+++ packages/types/src/accounts/ports/index.ts
@@ -1,8 +1,2 @@
 export { IAccountsPortName } from './IAccountsPort';
 export type { IAccountsPort } from './IAccountsPort';
-export { IMcpTokenServiceName } from './IMcpTokenService';
-export type {
-  IMcpTokenService,
-  McpTokenPayload,
-  McpTokenResponse,
-} from './IMcpTokenService';
